Fix x64 LuaJIT diagnostics
MSVC reports the mixed enum operands of the callback stubs (C5287) and
three narrowing assignments (C4244) at /W3. All four are intended, and the
engine build is kept free of warnings, so they get explicit casts.

Successor of ade7495d: its buildvm part only served the cleanup function of
the AddressSanitizer commit, which is not carried over.

+++ src/lj_ccallback.c
@@ -136,7 +136,7 @@ static void *callback_mcode_init(global_State *g, uint8_t *page)
     *(uint32_t *)p = XI_ENDBR64; p += 4;
 #endif
     /* mov al, slot; jmp group */
-    *p++ = XI_MOVrib | RID_EAX; *p++ = (uint8_t)slot;
+    *p++ = (uint8_t)((uint8_t)XI_MOVrib | (uint8_t)RID_EAX); *p++ = (uint8_t)slot;
     if ((slot & (CALLBACK_MCODE_NSLOT-1)) == (CALLBACK_MCODE_NSLOT-1) ||
 	slot == CALLBACK_MAX_SLOT-1) {
       /* push ebp/rbp; mov ah, slot>>8; mov ebp, &g. */
@@ -146,7 +146,7 @@ static void *callback_mcode_init(global_State *g, uint8_t *page)
       *p++ = 0x48; *p++ = XI_MOVri | RID_EBP;
       *(uint64_t *)p = (uint64_t)(g); p += 8;
 #else
-      *p++ = XI_MOVri | RID_EBP;
+      *p++ = (uint8_t)((uint8_t)XI_MOVri | (uint8_t)RID_EBP);
       *(int32_t *)p = i32ptr(g); p += 4;
 #endif
 #if LJ_64
